# wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "home-iot-api",
    "description" : "The API for the EatBacon IOT project",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/EDUARDOBANOSROMERO/Test/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  } ],
  "paths" : {
    "/devices" : {
      "get" : {
        "tags" : [ "Device" ],
        "description" : "returns all registered devices",
        "operationId" : "getDevices",
        "parameters" : [ {
          "name" : "skip",
          "in" : "query",
          "description" : "number of records to skip",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "limit",
          "in" : "query",
          "description" : "max number of records to return",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "All the devices",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string",
                    "format" : "uri",
                    "example" : "http://10.0.0.225:8080"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Device" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeviceRegistrationInfo"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "successfully registered device"
          }
        }
      }
    },
...
    "/temperature/forecast/{days}" : {
      "get" : {
        "tags" : [ "Environment" ],
        "operationId" : "getForecast",
        "parameters" : [ {
          "name" : "days",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "the forecast",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ForecastResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "Environment"
      }
    },
    "/temperature/{zoneId}" : {
      "get" : {
        "tags" : [ "Environment" ],
        "operationId" : "getZoneTemperature",
        "parameters" : [ {
          "name" : "zoneId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Zone temperature",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/TemperatueZoneStatus"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "Environment"
      }
    },
    "/temperature/{zoneId}/heater" : {
      "get" : {
        "tags" : [ "Environment" ],
        "description" : "gets the state of the heater",
        "operationId" : "getHeaterState",
        "parameters" : [ {
          "name" : "zoneId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "heater state",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/HeaterState"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "Environment"
      }
    },
    "/temperature/{zoneId}/heater/{state}" : {
      "post" : {
        "tags" : [ "Environment" ],
        "description" : "turns the heater on or off",
        "operationId" : "setHeaterState",
        "parameters" : [ {
          "name" : "zoneId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "state",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "false", "true" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Status of the operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          }
        },
        "x-swagger-router-controller" : "Environment"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "TemperatueZoneStatus" : {
        "required" : [ "id", "timestamp", "value" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "description" : "the unique identifier for the zone"
          },
          "name" : {
            "type" : "string",
            "description" : "the name of the zone"
          },
          "value" : {
            "type" : "number",
            "description" : "the temperature in the zone",
            "format" : "double"
          },
          "units" : {
            "type" : "string",
            "description" : "the temperature units",
            "enum" : [ "celsius", "fahrenheit" ],
            "default" : "fahrenheit"
          },
          "timestamp" : {
            "type" : "string",
            "description" : "the timestamp when the temperature was measured",
            "format" : "date-time"
          }
        },
        "description" : "status of a single zone",
        "example" : {
          "name" : "name",
          "id" : "id",
          "units" : "fahrenheit",
          "value" : 5.962133916683182,
          "timestamp" : "2000-01-23T04:56:07.000+00:00"
        }
      },
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "message" : {
            "type" : "string",
            "example" : "everything is ok"
          }
        },
        "example" : {
          "code" : 0,
          "message" : "everything is ok"
        }
      },
      "HeaterState" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          }
        },
        "example" : {
          "id" : "id",
          "state" : "state"
        }
      },
...
      "ForecastResponse" : {
        "type" : "object",
        "properties" : {
          "city" : {
            "$ref" : "#/components/schemas/City"
          },
          "values" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Forecast"
            }
          }
        },
        "example" : {
          "city" : {
            "country" : "country",
            "name" : "name",
            "lon" : 1.4658129805029452,
            "id" : 0,
            "lat" : 6.027456183070403
          },
          "values" : [ {
            "date" : "2000-01-23T04:56:07.000+00:00",
            "temperature" : {
              "high" : 3.616076749251911,
              "low" : 9.301444243932576,
              "night" : 1.2315135367772556,
              "evening" : 7.386281948385884,
              "day" : 4.145608029883936,
              "morning" : 2.027123023002322
            },
            "weather" : {
              "summary" : "summary",
              "icon" : "icon",
              "description" : "description"
            },
            "humidity" : 5,
            "pressure" : 5.962133916683182,
            "clouds" : 7,
            "windSpeed" : 2.3021358869347655
          }, {
            "date" : "2000-01-23T04:56:07.000+00:00",
            "temperature" : {
              "high" : 3.616076749251911,
              "low" : 9.301444243932576,
              "night" : 1.2315135367772556,
              "evening" : 7.386281948385884,
              "day" : 4.145608029883936,
              "morning" : 2.027123023002322
            },
            "weather" : {
              "summary" : "summary",
              "icon" : "icon",
              "description" : "description"
            },
            "humidity" : 5,
            "pressure" : 5.962133916683182,
            "clouds" : 7,
            "windSpeed" : 2.3021358869347655
          } ]
        }
      },
      "Forecast" : {
        "type" : "object",
        "properties" : {
          "date" : {
            "type" : "string",
            "format" : "date-time"
          },
          "pressure" : {
            "type" : "number",
            "format" : "double"
          },
          "humidity" : {
            "type" : "integer",
            "format" : "int32"
          },
          "windSpeed" : {
            "type" : "number",
            "format" : "double"
          },
          "clouds" : {
            "type" : "integer",
            "format" : "int32"
          },
          "temperature" : {
            "$ref" : "#/components/schemas/ForecastTemperature"
          },
          "weather" : {
            "$ref" : "#/components/schemas/WeatherForecast"
          }
        },
        "example" : {
          "date" : "2000-01-23T04:56:07.000+00:00",
          "temperature" : {
            "high" : 3.616076749251911,
            "low" : 9.301444243932576,
            "night" : 1.2315135367772556,
            "evening" : 7.386281948385884,
            "day" : 4.145608029883936,
            "morning" : 2.027123023002322
          },
          "weather" : {
            "summary" : "summary",
            "icon" : "icon",
            "description" : "description"
          },
          "humidity" : 5,
          "pressure" : 5.962133916683182,
          "clouds" : 7,
          "windSpeed" : 2.3021358869347655
        }
      },
      "City" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "name" : {
            "type" : "string"
          },
          "lat" : {
            "type" : "number",
            "format" : "double"
          },
          "lon" : {
            "type" : "number",
            "format" : "double"
          },
          "country" : {
            "type" : "string"
          }
        },
        "example" : {
          "country" : "country",
          "name" : "name",
          "lon" : 1.4658129805029452,
          "id" : 0,
          "lat" : 6.027456183070403
        }
      },
      "ForecastTemperature" : {
        "type" : "object",
        "properties" : {
          "low" : {
            "type" : "number",
            "format" : "double"
          },
          "high" : {
            "type" : "number",
            "format" : "double"
          },
          "morning" : {
            "type" : "number",
            "format" : "double"
          },
          "day" : {
            "type" : "number",
            "format" : "double"
          },
          "evening" : {
            "type" : "number",
            "format" : "double"
          },
          "night" : {
            "type" : "number",
            "format" : "double"
          }
        },
        "example" : {
          "high" : 3.616076749251911,
          "low" : 9.301444243932576,
          "night" : 1.2315135367772556,
          "evening" : 7.386281948385884,
          "day" : 4.145608029883936,
          "morning" : 2.027123023002322
        }
      },
      "WeatherForecast" : {
        "type" : "object",
        "properties" : {
          "summary" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "icon" : {
            "type" : "string"
          }
        },
        "example" : {
          "summary" : "summary",
          "icon" : "icon",
          "description" : "description"
        }
      },
      "DeviceRegistrationInfo" : {
        "type" : "object",
        "properties" : {
          "uri" : {
            "type" : "string",
            "format" : "uri",
            "example" : "http://10.0.0.220:8080"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid",
            "example" : "0729a580-2240-11e6-9eb5-0002a5d5c51b"
          }
        }
      }
    }
  }
}